Move modules transpilation to calypso-build
