+++ webpack.config.js
@@ -19,7 +19,10 @@ const InlineConstantExportsPlugin = require( '@automattic/webpack-inline-constan
 const Minify = require( '@automattic/calypso-build/webpack/minify' );
 const SassConfig = require( '@automattic/calypso-build/webpack/sass' );
 const TranspileConfig = require( '@automattic/calypso-build/webpack/transpile' );
-const { cssNameFromFilename } = require( '@automattic/calypso-build/webpack/util' );
+const {
+	cssNameFromFilename,
+	shouldTranspileDependency,
+} = require( '@automattic/calypso-build/webpack/util' );
 
 /**
  * Internal dependencies
@@ -99,42 +102,6 @@ function createProgressHandler() {
 
 		console.log( message ); // eslint-disable-line no-console
 	};
-}
-
-const nodeModulesToTranspile = [
-	// general form is <package-name>/.
-	// The trailing slash makes sure we're not matching these as prefixes
-	// In some cases we do want prefix style matching (lodash. for lodash.assign)
-	'd3-array/',
-	'd3-scale/',
-	'debug/',
-];
-/**
- * Check to see if we should transpile certain files in node_modules
- * @param {String} filepath the path of the file to check
- * @returns {Boolean} True if we should transpile it, false if not
- *
- * We had a thought to try to find the package.json and use the engines property
- * to determine what we should transpile, but not all libraries set engines properly
- * (see d3-array@2.0.0). Instead, we transpile libraries we know to have dropped Node 4 support
- * are likely to remain so going forward.
- */
-function shouldTranspileDependency( filepath ) {
-	// find the last index of node_modules and check from there
-	// we want <working>/node_modules/a-package/node_modules/foo/index.js to only match foo, not a-package
-	const marker = '/node_modules/';
-	const lastIndex = filepath.lastIndexOf( marker );
-	if ( lastIndex === -1 ) {
-		// we're not in node_modules
-		return false;
-	}
-
-	const checkFrom = lastIndex + marker.length;
-
-	return nodeModulesToTranspile.some(
-		nodeModulesToTranspile,
-		modulePart => filepath.substring( checkFrom, checkFrom + modulePart.length ) === modulePart
-	);
 }
 
 let outputFilename = '[name].[chunkhash].min.js'; // prefer the chunkhash, which depends on the chunk, not the entire build
